chore - move the oven command options and store helpers beside the commands (#1299)
`src/cli/commands/oven.rs` carries ~5,900 production lines, most of it the
compiler-suite run and its helpers. Two things in there are not that, and both
have obvious edges:

- `options` -- one struct per `incan oven` command surface, carrying what the CLI
  parsed and no behavior beyond a small accessor.
- `support` -- opening a store with the right defaults, resolving byte limits
  from the environment, and the formatting and error-wrapping helpers the command
  implementations reach for.

Moving them leaves the command implementations reading as implementations. No
logic is edited.

Items that were module-private are now `pub(super)`, which is the reach they had
when they shared a file with their callers; nothing became visible outside this
module. Imports each block needs are declared where it lives.

Evidence it is a move: the exported surface is identical -- 31 items before and
after, none lost, none gained -- and `cargo test --lib` reports the same
3234 passed / 28 failed as the unmodified branch, those 28 being the known
`CARGO_BIN_EXE_incan` gap every plain `--lib` run hits.

This completes the three files #1299 names, one per pull request. The
compiler-suite cluster that dominates this file carries behavior and is worth
splitting against its own evidence rather than alongside option shapes.
